Add flag to enable/disable creation of resources
When this module is used with terraform workspaces we may choose to
enable/disable creation of resources based on various conditions.  Since
terraform modules don't support the count attribute we must pass in the
parameter and apply it to each resource within the module.

Unfortunately we also need to perform magic around various references to
paramaters for each resource when it doesn't get created.  We do this by
applying defaults with a combination of interpolation functions in a
very hacky way.  Not something to be proud of.