Add account ID to GuardDuty findings
In order to add this attribute, I validated that even on a single (newly created) AWS account, the GuardDuty event always contains an `AccountID` attribute.

The AWS documentation https://docs.aws.amazon.com/guardduty/latest/ug/guardduty_findings_cloudwatch.html seems to indicate this too.

I've modified the sample events and messages to add a bogus `accountId`. Running the `pipenv` commands as per the `README` made some formatting changes on `mylambda.py`, so I am leaving that as suggested by that run.

The snapshots also needed updating to reflect the fact that we now have an AccountID in the GuardDuty event/message.

Refs waysact/devops#9087